Package edits for generating one swagger per resource type
TODO: do the same for FHIROpenApiGenerator and try out the OpenAPI files
in liberty.
